feat: Professional Vocal Analysis System v2 - 전문가급 보컬 분석 시스템 완성
✨ 새로운 기능
- Virtual Listener: 다운로드 없이 YouTube 30초 스트리밍 분석
- Professional Vocal Analyzer: 7가지 발성 구역 분류 시스템
- SQLite 라벨링 데이터베이스 영구 저장 시스템
- 교육학적 평가 엔진 (100점 만점, A+ ~ D 등급)

🎼 전문 분석 기능
- 발성 레지스터: Vocal Fry, Chest, Mixed, Head, Falsetto, Whistle, Passaggio
- IPA 기반 음성학적 모음 분석
- 전문가 수준 비브라토 분석 (속도 4-7Hz, 깊이 0-200센트)
- 포먼트 기반 공명 패턴 분석
- 음성 건강 모니터링 및 strain 감지

🏥 건강 중심 접근법
- 실시간 성대 피로도 모니터링
- 호흡 효율성 평가
- 위험 수준 분류 (Safe/Caution/Risk)
- 예방 중심 권장사항

📊 데이터베이스
- vocal_labels.db 자동 생성
- 배치 분석 + 자동 DB 저장
- JSON 내보내기 지원

🐛 버그 수정
- TextField 입력 문제 해결 (FocusNode 관리)
- SQLite 스레드 오류 수정 (check_same_thread=False)
- Flutter UI 컴포넌트 Material 래핑 추가

🔗 API 엔드포인트
- POST /professional_analysis - 전문가급 종합 분석
- GET /vocal_health_report - 음성 건강 리포트
- GET /technique_analysis - 발성 기법 분석
- GET /learning_progress - 학습 진도 추적
- GET /pedagogical_assessment/<id> - 교육학적 평가

+++ lib/core/advanced_pitch_tracker.dart
@@ -284,7 +284,7 @@ class AdvancedPitchTracker {
     
     return PitchData(
       frequency: frequency,
-      confidence: confidence.clamp(0, 1),
+      confidence: confidence.toDouble().clamp(0.0, 1.0),
       timestamp: DateTime.now(),
     );
   }
